Moving withTags to the plugin api, adding perms to Feature comments
