+++ plugin-api/beta/client/hocs/index.js
@@ -1,1 +1,2 @@
 export {default as withReaction} from './withReaction';
+export {default as withTags} from './withTags';
