+++ plugin-api/beta/client/hocs/withTags.js
