+++ plugins/talk-plugin-featured-comments/client/components/FeaturedButton.js
@@ -1,15 +1,15 @@
 import React from 'react';
-import withTags from './withTags';
 import styles from './styles.css';
 import cn from 'classnames';
-import {t} from 'plugin-api/beta/client/services';
+import {t, can} from 'plugin-api/beta/client/services';
 import {name} from '../../package.json';
+import {withTags} from 'plugin-api/beta/client/hocs';
 import {Icon} from 'plugin-api/beta/client/components/ui';
 
 const FeaturedButton = (props) => {
-  const {alreadyTagged, deleteTag, postTag} = props;
+  const {alreadyTagged, deleteTag, postTag, user} = props;
 
-  return (
+  return can(user, 'MODERATE_COMMENTS') ? (
     <button 
       className={cn([name, styles.button, {[styles.featured] : alreadyTagged}])}
       onClick={alreadyTagged ? deleteTag : postTag} >
@@ -25,7 +25,7 @@ const FeaturedButton = (props) => {
       )}
 
     </button>
-  );
+  ) : null ;
 };
 
 export default withTags('featured')(FeaturedButton);
